feat: implement reservation pricing engine with guest details and enum-based room type handling
- Added guestName, guestEmail, guestPhone to reservations
- Replaced username with createdBy for proper audit tracking
- Implemented totalPrice calculation using room pricing and nights
- Integrated RoomPricing repository for dynamic pricing
- Ensured type safety using enum-based roomType mapping
- Fixed schema alignment and removed legacy room_id references
- Improved reservation response structure for real-world use cases

+++ src/main/java/com/example/HRMS/controller/AnalyticsController.java
@@ -5,6 +5,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.security.access.prepost.PreAuthorize;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 @RestController
@@ -16,7 +17,7 @@ public class AnalyticsController {
 
     @PreAuthorize("hasRole('ADMIN')")
     @GetMapping("/revenue")
-    public Double getRevenue() {
+    public BigDecimal getRevenue() {
         return analyticsService.getTotalRevenue();
     }
 
@@ -28,7 +29,7 @@ public class AnalyticsController {
 
     @PreAuthorize("hasRole('ADMIN')")
     @GetMapping("/monthly-revenue")
-    public Map<String, Double> getMonthlyRevenue() {
+    public Map<String, BigDecimal> getMonthlyRevenue() {
         return analyticsService.getMonthlyRevenue();
     }
